gearstore_backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
drf-nested-routers==0.94.1
python-dotenv==1.0.1
requests==2.32.3
